Update unit test and improve search ranking
This commit updates the unit test in the IndexerTest class to expect one hit instead of two. It also improves the search ranking algorithm in the FuzzySearch class by using a weighted ratio instead of a partial ratio. This will produce more relevant search results by placing higher weight on the order and proximity of character matches.